+++ CHANGELOG.md
@@ -49,6 +49,7 @@
   value, because `CancelToken` is not `Copy`.
 - Make Rhai `vec2(..)` and `vec3(..)` functions idempotent, i.e.
   `vec2(vec2(..))` now works.
+- Optimize `atan2` interval evaluation
 
 # 0.3.8
 - Bug fix: `Image::height()` was returning width instead!